% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caseControlGE.R
\docType{package}
\name{caseControlGE-package}
\alias{caseControlGE}
\alias{caseControlGE-package}
\title{Semiparametric Gene-Environment Interactions in Case-Control Studies}
\description{
This package was written to perform simulations for the follow up to Stalder et. al. (2017).
}
\details{
\code{\link{simfun}} Simulates then maximizes the retrospective pseudolikelihood of case-control data under the
assumption of G-E independence.  It returns estimates, asymptotic standard error, and
(optionally) bootstrapped standard error for several versions of the estimator.

\code{\link{simulate_complex}} simulates data to be analyzed by \code{simfun},
logistic regression, or other models.  This function was used to simulate data for
Stalder et. al. (2017).  The user can specify up to three types of genetic variables
(SNPs with additive effects under HWE, normally distributed polygenic risk scores,
and gamma distributed polygenic risk scores), each of which can be multivariate.
Two types of environmental variables (binary and normal) can also be potentially multivariate.

\code{\link{combo_asymp}} is called by \code{\link{simfun}} to calculate estimates and asymptotic
SE.

\code{\link{combo_boot}} is called by \code{\link{simfun}} to calculate bootstrap estimates.  It
is called after \code{\link{combo_asymp}}
}
