% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_functions.R
\name{combo_boot}
\alias{combo_boot}
\title{Function to calculate bootstrapped SE for SPMLE_G & SPMLE_E estimates, as well as bootstrapped Combo estimate}
\usage{
combo_boot(D, G, E, pi1, SPMLE_G_par, SPMLE_E_par, control = list(),
  current_sim = NULL, logpath, filename)
}
\arguments{
\item{D}{a binary vector of disease status (1=case, 0=control).}

\item{G}{a vector or matrix (if multivariate) containing genetic data. Can be continuous, discrete, or a combination.}

\item{E}{a vector or matrix (if multivariate) containing environmental data. Can be continuous, discrete, or a combination.}

\item{pi1}{the population disease rate, a scalar in [0, 1).  Using \code{pi1=0} is the rare disease approximation.}

\item{SPMLE_G_par, SPMLE_E_par}{Parameter estimates generated from \code{\link{combo_asymp}}}

\item{control}{a list of control parameters.
\code{nboot} number of bootstraps.  Default \code{0}.
\code{trace} is a scalar.  If >-1, tracing information is produced.  Default \code{0}.
\code{usehess} logical: precondition optimization with the hessian.  Default \code{TRUE}.}

\item{current_sim}{Name or number of current simulation.  Default \code{NULL}.}

\item{logpath}{paths for saving files}

\item{filename}{file name for log and simulations}
}
\value{
A list with a matrix of estimates and SEs and a bunch of other stuff
}
\description{
Function to calculate bootstrapped SE for SPMLE_G & SPMLE_E estimates, as well as bootstrapped Combo estimate
}
\seealso{
\code{\link{simulate_complex}} to simulate data
}
