% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_compare.R
\name{compare_fun}
\alias{compare_fun}
\title{Function to compare the performance of the various estimators}
\usage{
compare_fun(sim_df, truevec = NULL, all_ests = TRUE, pi1 = NULL,
  ncase = NULL, ncontrol = NULL, beta0 = NULL, kappa = NULL,
  betaG_SNP = NULL, betaG_normPRS = NULL, betaG_gammaPRS = NULL,
  betaG_bimodalPRS = NULL, betaE_bin = NULL, betaE_norm = NULL,
  betaGE_SNP_bin = NULL, betaGE_normPRS_bin = NULL,
  betaGE_gammaPRS_bin = NULL, betaGE_bimodalPRS_bin = NULL,
  betaGE_SNP_norm = NULL, betaGE_normPRS_norm = NULL,
  betaGE_gammaPRS_norm = NULL, betaGE_bimodalPRS_norm = NULL)
}
\arguments{
\item{sim_df}{data frame of simulation results}

\item{truevec}{true parameter values, with kappa as the intercept.  User must provide truevec OR separate betas.  Default \code{NULL}.}

\item{all_ests}{Logical: include all estimates. If \code{FALSE}, leave out simple combo, weighted combo, and combo with bootstrapped Lambda. Default \code{TRUE}.}

\item{pi1}{population disease rate.  Default \code{NULL}.}

\item{ncase, ncontrol}{number of cases and controls.  Default \code{NULL}.}

\item{beta0}{logistic intercept.  Default \code{NULL}.}

\item{kappa}{intercept estimated by logistic regression.  If not provided, will be calculated from pi1, ncase, ncontrol, and beta0.  Default \code{NULL}.}

\item{betaG_SNP, betaG_normPRS, betaG_gammaPRS, betaG_bimodalPRS}{coefficients for genetic variables.  Default \code{NULL}.}

\item{betaE_bin, betaE_norm}{coefficients for environmental variables.  Default \code{NULL}.}

\item{betaGE_SNP_bin, betaGE_normPRS_bin, betaGE_gammaPRS_bin, betaGE_bimodalPRS_bin, betaGE_SNP_norm, betaGE_normPRS_norm, betaGE_gammaPRS_norm, betaGE_bimodalPRS_norm}{coefficients for G-E interactions.  Default \code{NULL}.}
}
\value{
a data frame with results of the comparison.  If \code{all_ests=TRUE}, attributes will include a list with coverage matrices for the combination (fusion) estimators.
}
\description{
Function to compare the performance of the various estimators
}
