% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{enn}
\alias{enn}
\title{Checks whether a variable exists and is not NULL}
\usage{
enn(varname)
}
\arguments{
\item{varname}{variable name}
}
\value{
TRUE or FALSE
}
\description{
Checks whether a variable exists and is not NULL
}
