% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateCC.R
\name{lenn}
\alias{lenn}
\title{Returns the length of a variable (0 if missing or null) whether a variable exists (in the current environment) and is not NULL}
\usage{
lenn(varname)
}
\arguments{
\item{varname}{variable name}
}
\value{
length of varname
}
\description{
Returns the length of a variable (0 if missing or null) whether a variable exists (in the current environment) and is not NULL
}
