% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_compare.R
\name{read_matrices}
\alias{read_matrices}
\title{Function to read saved matrices}
\usage{
read_matrices(filename, simpath = "./sims/", matrix = "asymp_ests$Sigma_GE")
}
\arguments{
\item{filename}{file name prefix}

\item{simpath}{path to find simulations}

\item{matrix}{matrix to read}
}
\value{
a list of matrices
}
\description{
Function to read saved matrices
}
