% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_compare.R
\name{read_sims}
\alias{read_sims}
\title{Function to read saved simulations}
\usage{
read_sims(filename, simpath = "./sims/")
}
\arguments{
\item{filename}{file name prefix}

\item{simpath}{path to find simulations}
}
\value{
A data frame of simulation results
}
\description{
Function to read saved simulations
}
