% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfun.R
\name{simfun}
\alias{simfun}
\title{Simulate data and run the analysis}
\usage{
simfun(X, control = list(), pi1, ncase, ncontrol, beta0, betaG_SNP,
  betaG_normPRS, betaG_gammaPRS, betaG_bimodalPRS, betaE_bin, betaE_norm,
  betaGE_SNP_bin, betaGE_normPRS_bin, betaGE_gammaPRS_bin,
  betaGE_bimodalPRS_bin, betaGE_SNP_norm, betaGE_normPRS_norm,
  betaGE_gammaPRS_norm, betaGE_bimodalPRS_norm, MAF, SNP_cor, G_normPRS_cor,
  E_bin_freq, E_norm_cor, regress_E_bin_on_G_SNP, regress_E_bin_on_G_normPRS,
  regress_E_bin_on_G_gammaPRS, regress_E_bin_on_G_bimodalPRS,
  regress_E_norm_on_G_SNP, regress_E_norm_on_G_normPRS,
  regress_E_norm_on_G_gammaPRS, regress_E_norm_on_G_bimodalPRS)
}
\arguments{
\item{X}{simulation number}

\item{control}{a list of control parameters.
\code{nboot} number of bootstraps.  Default \code{0}.
\code{trace} is a scalar.  If >-1, tracing information is produced.  Default \code{0}.
\code{usehess} logical: precondition optimization with the hessian.  Default \code{TRUE}.
\code{useseed} logical: use the simulation number as the seed.  Default \code{TRUE}.}

\item{pi1}{pipulation disease rate}

\item{ncase, ncontrol}{number of cases and controls}

\item{beta0}{logistic intercept}

\item{betaG_SNP, betaG_normPRS, betaG_gammaPRS, betaG_bimodalPRS}{coefficients for genetic variables}

\item{betaE_bin, betaE_norm}{coefficients for environmental variables}

\item{betaGE_SNP_bin, betaGE_normPRS_bin, betaGE_gammaPRS_bin, betaGE_bimodalPRS_bin, betaGE_SNP_norm, betaGE_normPRS_norm, betaGE_gammaPRS_norm, betaGE_bimodalPRS_norm}{coefficients for G-E interactions}

\item{MAF}{Mean Allele Frequency}

\item{SNP_cor}{correlation between successive SNPs}

\item{G_normPRS_cor}{correlation between multivariate normal genetic PRS variables}

\item{E_bin_freq}{frequency of E=1 for binary environmental variables}

\item{E_norm_cor}{correlation between multivariate normal environmental variables}

\item{regress_E_bin_on_G_SNP, regress_E_bin_on_G_normPRS, regress_E_bin_on_G_gammaPRS, regress_E_bin_on_G_bimodalPRS, regress_E_norm_on_G_SNP, regress_E_norm_on_G_normPRS, regress_E_norm_on_G_gammaPRS, regress_E_norm_on_G_bimodalPRS}{to violate the G-E independence assumption}
}
\value{
A named row vector of estimates and SEs
}
\description{
Simulate data and run the analysis
}
