% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_complex.R
\name{simulate_complex}
\alias{simulate_complex}
\title{Simulate case-control data with multivariate, possibly dependent genetic and environmental components.}
\usage{
simulate_complex(ncase, ncontrol, beta0, betaG_SNP = NULL,
  betaG_normPRS = NULL, betaG_gammaPRS = NULL, betaG_bimodalPRS = NULL,
  betaE_bin = NULL, betaE_norm = NULL, betaGE_SNP_bin = NULL,
  betaGE_normPRS_bin = NULL, betaGE_gammaPRS_bin = NULL,
  betaGE_bimodalPRS_bin = NULL, betaGE_SNP_norm = NULL,
  betaGE_normPRS_norm = NULL, betaGE_gammaPRS_norm = NULL,
  betaGE_bimodalPRS_norm = NULL, MAF = NULL, SNP_cor = 0,
  G_normPRS_cor = 0, E_bin_freq = 0.5, E_norm_cor = 0,
  regress_E_bin_on_G_SNP = NULL, regress_E_bin_on_G_normPRS = NULL,
  regress_E_bin_on_G_gammaPRS = NULL, regress_E_bin_on_G_bimodalPRS = NULL,
  regress_E_norm_on_G_SNP = NULL, regress_E_norm_on_G_normPRS = NULL,
  regress_E_norm_on_G_gammaPRS = NULL,
  regress_E_norm_on_G_bimodalPRS = NULL, control = list())
}
\arguments{
\item{ncase, ncontrol}{number of cases and controls.}

\item{beta0}{logistic intercept. Can be manipulated to change the disease rate.}

\item{betaG_SNP, betaG_normPRS, betaG_gammaPRS, betaG_bimodalPRS}{coefficients for genetic variable main effects.
Genetic variables can include SNPs and polygenic risk scores with normal, gamma, and bimodal distributions.
A value of \code{NULL} for coefficient of a genetic variable means that type of genetic variable will not be generated.
Vector valued coefficients produce multivariate genetic data.}

\item{betaE_bin, betaE_norm}{coefficients for environmental variable main effects. Environmental variables can include
binary and normally distributed random variables. A value of \code{NULL} for coefficient of an environmental variable
means that type of environmental variable will not be generated. Vector valued coefficients produce multivariate environmental data.}

\item{betaGE_SNP_bin, betaGE_normPRS_bin, betaGE_gammaPRS_bin, betaGE_bimodalPRS_bin, betaGE_SNP_norm, betaGE_normPRS_norm, betaGE_gammaPRS_norm, betaGE_bimodalPRS_norm}{coefficients for
multiplicative G*E effects. The length of the coefficient of any given G*E interaction must be equal to the product of the lengths
of the coefficients of the corresponding G and E main effects. If one of the main effects is \code{NULL} the interaction must be \code{NULL} as well.}

\item{MAF}{Minor Allele Frequency of SNPs. This vector is the same length as \code{beta_G_SNP} and has values between 0 and 1.
The MAF is used to generate SNP data that is in Hardy-Weinberg Equilibrium.}

\item{SNP_cor}{scalar between values between -1 and 1. SNPs are simulated by generating multivariate normal random draws with an AR(\code{SNP_cor})
covariance matrix. These normal draws are then trichotomized according to HWE to simulate SNPs.}

\item{G_normPRS_cor}{correlation matrix for multivariate normal polygenic risk scores. In the bivariate case, a 2x2 matrix or a scalar
(for correlation) are accepted. NULL values indicate independent polygenic risk scores.}

\item{E_bin_freq}{marginal probability that E_bin = 1. Must have length equal to \code{length(betaE_bin)} and values between 0 and 1.}

\item{E_norm_cor}{correlation matrix for multivariate normal environmental variable. In the bivariate case, a 2x2 matrix or a scalar
(for correlation) are accepted. NULL values indicate independent normal environmental variables.}

\item{regress_E_bin_on_G_SNP, regress_E_bin_on_G_normPRS, regress_E_bin_on_G_gammaPRS, regress_E_bin_on_G_bimodalPRS}{allow the simulation of case-control data that violates
the G-E independence assumption. If these arguments are \code{NULL} or all \code{0}s, the binary environmental variables will be independent
of the genetic variables. If non-null, the length of the regression argument must be equal to the product of the lengths of the coefficients
of the corresponding G and E main effects. The conditional expectations of binary environmental variables depend on the product of the
\code{regress_E_bin} arguments and the corresponding genetic variables.}

\item{regress_E_norm_on_G_SNP, regress_E_norm_on_G_normPRS, regress_E_norm_on_G_gammaPRS, regress_E_norm_on_G_bimodalPRS}{allow the simulation of case-control data that violates
the G-E independence assumption. If these arguments are \code{NULL} or all \code{0}s, the normal environmental variables will be independent
of the genetic variables. If non-null, the length of the regression argument must be equal to the product of the lengths of the coefficients
of the corresponding G and E main effects. The conditional means of normal environmental variables depend on the product of the
\code{regress_E_norm} arguments and the corresponding genetic variables.}

\item{control}{a list of control parameters.
\code{trace} is a scalar.  If >-1, tracing information is produced.  Default \code{trace=0}.}
}
\value{
\code{simulate_complex} produces a list with three elements:
 \item{\code{D}}{a binary vector with \code{ncontrol} \code{0}s and \code{ncase} \code{1}s.}
 \item{\code{G}}{a matrix with \code{ncontrol + ncase} rows and a column for each genetic variable. Genetic variables are ordered: SNPs, normal PRSs, gamma PRSs.}
 \item{\code{E}}{a matrix with \code{ncontrol + ncase} rows and a column for each environmental variable. Environmental variables are ordered: binary, normal.}
}
\description{
\code{simulate_complex} simulates data to be analyzed by \code{symple}, logistic regression, or other models.
}
\details{
This function was used to simulate data for Stalder et. al. (2017).
The user can specify up to four types of genetic variables (SNPs with additive effects under HWE,
polygenic risk scores with normal, gamma, and bimodal distributions), each of which can be multivariate.
Two types of environmental variables (binary and normal) can also be potentially multivariate.

If both G and E are multivariate, \code{beta_GE_} and \code{regress_E_} arguments iterate G quickly and E slowly.
For example, if G_SNP and E_bin are both bivariate, \code{betaGE_SNP_bin} is ordered (G1*E1, G2*E1, G1*E2, G2*E2).
}
\examples{
# Simulation from Table 1 in Stalder et. al. (2017)
dat = simulate_complex(ncase=1000,
                       ncontrol=1000,
                       beta0=-4.14,
                       betaG_SNP=c(log(1.2), log(1.2), 0, log(1.2), 0),
                       betaE_bin=log(1.5),
                       betaGE_SNP_bin=c(log(1.3), 0, 0, log(1.3), 0),
                       MAF=c(0.1, 0.3, 0.3, 0.3, 0.1),
                       SNP_cor=0.7,
                       E_bin_freq=0.5)

# Simulation with 5 SNPs and a single normal environmental variable
# that is dependent on G1 with an R^2 of 0.001.
# True population disease rate in this simulation is 0.03.
# This simulation scenario was used in the Supplementary Material of Stalder et. al. (2017)
dat = simulate_complex(ncase=1000,
                       ncontrol=1000,
                       beta0=-4.19,
                       betaG_SNP=c(log(1.2), log(1.2), 0, log(1.2), 0),
                       betaE_norm=(qnorm(0.75)-qnorm(0.25))*c(log(1.5)),
                       betaGE_SNP_norm=(qnorm(0.75)-qnorm(0.25))*c(log(1.3), 0, 0, log(1.3), 0),
                       MAF=c(0.1, 0.3, 0.3, 0.3, 0.1),
                       SNP_cor=0.7,
                       regress_E_norm_on_G_SNP=c(sqrt(0.001),rep(0,4)),
                       control=list(trace=1))
}
\seealso{
\code{symple}
}
