% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateCC.R
\name{simulateCC}
\alias{simulateCC}
\title{Simulate case-control data with multivariate, possibly dependent genetic and environmental components.}
\usage{
simulateCC(ncase, ncontrol, beta0, betaG_SNP, betaG_normPRS, betaG_gammaPRS,
  betaG_bimodalPRS, betaE_bin, betaE_norm, betaGE_SNP_bin, betaGE_normPRS_bin,
  betaGE_gammaPRS_bin, betaGE_bimodalPRS_bin, betaGE_SNP_norm,
  betaGE_normPRS_norm, betaGE_gammaPRS_norm, betaGE_bimodalPRS_norm, MAF,
  SNP_cor = 0, G_normPRS_cor = 0, E_norm_cor = 0, E_bin_freq,
  regress_E_bin_on_G_SNP, regress_E_bin_on_G_normPRS,
  regress_E_bin_on_G_gammaPRS, regress_E_bin_on_G_bimodalPRS,
  regress_E_norm_on_G_SNP, regress_E_norm_on_G_normPRS,
  regress_E_norm_on_G_gammaPRS, regress_E_norm_on_G_bimodalPRS,
  control = list())
}
\arguments{
\item{ncase, ncontrol}{number of cases and controls, both must be positive integers.}

\item{beta0}{logistic intercept, required. Can be manipulated to change the population disease rate.}

\item{betaG_SNP, betaG_normPRS, betaG_gammaPRS, betaG_bimodalPRS}{optional coefficients for
genetic main effects (at least one must be specified).  Genetic variables can include SNPs and polygenic
risk scores with normal, gamma, and bimodal distributions.  Vector valued coefficients produce multivariate genetic data.
When simulating SNPs, you must provide \code{MAF} with the same length as \code{betaG_SNP}.}

\item{betaE_bin, betaE_norm}{optional coefficients for environmental variable main effects (at least one must be specified).
Environmental variables can include binary and normally distributed random variables.
Vector valued coefficients produce multivariate environmental data.}

\item{betaGE_SNP_bin, betaGE_normPRS_bin, betaGE_gammaPRS_bin, betaGE_bimodalPRS_bin}{,}

\item{betaGE_SNP_norm, betaGE_normPRS_norm, betaGE_gammaPRS_norm, betaGE_bimodalPRS_norm}{coefficients for
multiplicative G*E interaction effects.  The length of the coefficient of any
given G*E interaction must equal the product of the lengths
of the coefficients of the corresponding G and E main effects.}

\item{MAF}{Minor Allele Frequency of SNPs. This vector is the same length as
\code{beta_G_SNP} and has values between 0 and 1.
The MAF is used to generate SNP data that is in Hardy-Weinberg Equilibrium.
This specifies \code{Pr[G=(0, 1, 2)] =} \code{[(1-MAF)^2, 2*MAF(1-MAF), MAF^2]}.}

\item{SNP_cor}{scalar specifying the correlation between adjacent SNPs.
SNPs are simulated by generating multivariate normal random draws with an AR1(\code{SNP_cor})
covariance matrix. These normal draws are then trichotomized according to HWE to simulate SNPs.  Default \code{SNP_cor = 0}.}

\item{G_normPRS_cor}{correlation matrix for multivariate normal polygenic risk scores. In the bivariate case, a 2x2 matrix or a scalar
(for correlation) are accepted.  Default \code{G_normPRS_cor = 0} generates independent normal polygenic risk scores.}

\item{E_norm_cor}{correlation matrix for multivariate normal environmental variable. In the bivariate case, a 2x2 matrix or a scalar
(for correlation) are accepted.  Default \code{E_norm_cor = 0} generates independent normal environmental variables.}

\item{E_bin_freq}{marginal probability that E_bin = 1. Must have length equal to \code{length(betaE_bin)} and values between 0 and 1.}

\item{regress_E_bin_on_G_SNP, regress_E_bin_on_G_normPRS, regress_E_bin_on_G_gammaPRS}{,}

\item{regress_E_bin_on_G_bimodalPRS}{allow the simulation of case-control data that violates
the G-E independence assumption.  If specified, binary environmental variables
will be generated with \code{Pr(E=1|G) =} \code{plogis[regress_E_bin_on_G * G +}
\code{qlogis(E_bin_freq)]}.  If these arguments are missing, \code{NULL},
or all \code{0}s, the binary environmental variables will be independent
of the genetic variables.  If specified, the length of the regression argument
must equal the product of the lengths of the coefficients of the corresponding
G and E main effects.}

\item{regress_E_norm_on_G_SNP, regress_E_norm_on_G_normPRS, regress_E_norm_on_G_gammaPRS}{,}

\item{regress_E_norm_on_G_bimodalPRS}{allow the simulation of case-control data that violates
the G-E independence assumption.  If specified, normal environmental variables
will be generated from a Normal(\code{regress_E_norm_on_G * G}, 1) distribution.
If these arguments are missing, \code{NULL}, or all \code{0}s, the normal
environmental variables will be independent of the genetic variables.
If specified, the length of the regression argument must equal the product
of the lengths of the coefficients of the corresponding G and E main effects.}

\item{control}{a list of control parameters, all of which are ignored except
\code{trace}, a scalar.  If \code{trace > -1}, information about the simulation
(e.g. population disease rate, correlations between SNPs, etc.) is produced.
Default \code{trace=0}.}
}
\value{
\code{simulateCC} produces a list with three elements:
  \item{\code{D}}{a binary vector with \code{ncontrol} zeros and \code{ncase} ones.}
  \item{\code{G}}{a matrix with \code{ncontrol + ncase} rows and a column for
    each genetic variable. Genetic variables are ordered: SNPs, normal PRSs,
    gamma PRSs, bimodal PRSs.}
  \item{\code{E}}{a matrix with \code{ncontrol + ncase} rows and a column for
    each environmental variable. Environmental variables are ordered: binary, normal.}
}
\description{
\code{simulateCC} simulates case-control data to be analyzed by \code{\link{spmle}},
\code{\link{spmleCombo}}, logistic regression, or other methods.
}
\details{
The user can specify up to four types of genetic variables, each of which can
be multivariate: SNPs with additive effects under Hardy-Weinberg Equilibrium
and polygenic risk scores with normal, gamma, and bimodal distributions.
Two types of environmental variables (binary and normal) can also be potentially multivariate.

SNPs may be generated in linkage disequilibrium, yielding correlated SNPs.
Multivariate normal polygenic risk scores may have a user-specified correlation
matrix, as may multivariate normal environmental variables.  Correlation may
also be specified between genetic and environmental variables to simulate
data in violation of the gene-environment independence assumption.

The number of variables generated is determined by the length of the betas given.
If you specify \code{betaG_normPRS = c(log(1.1), log(1.2))} and
\code{betaE_bin = c(log(1.2), log(1.2), log(1.2))}, you will get two \code{G}
variables (normally distributed polygenic risk scores), and three \code{E}
variables (with binary distributions).  In this example, you would supply
a vector of length 6 for \code{betaGE_normPRS_bin}.

If both G and E are multivariate, \code{beta_GE_} and \code{regress_E_} arguments
iterate G quickly and E slowly.  In the example above, \code{betaGE_normPRS_bin}
is ordered (G1*E1, G2*E1, G1*E2, G2*E2, G1*E3, G2*E3).
}
\examples{
set.seed(2018)
# Simulation from Table 1 in Stalder et. al. (2017)
dat = simulateCC(ncase=1000,
                 ncontrol=1000,
                 beta0=-4.165,
                 betaG_SNP=c(log(1.2), log(1.2), 0, log(1.2), 0),
                 betaE_bin=log(1.5),
                 betaGE_SNP_bin=c(log(1.3), 0, 0, log(1.3), 0),
                 MAF=c(0.1, 0.3, 0.3, 0.3, 0.1),
                 SNP_cor=0.7,
                 E_bin_freq=0.5)

# Simulation with 5 SNPs and a single normal environmental variable
# that is dependent on G1 with an R^2 of 0.001.
# True population disease rate in this simulation is 0.03.
# This simulation scenario was used in the Supplementary Material of Stalder et. al. (2017)
dat2 = simulateCC(ncase=1000,
                  ncontrol=1000,
                  beta0=-3.89,
                  betaG_SNP=c(log(1.2), log(1.2), 0, log(1.2), 0),
                  betaE_norm=log(1.5)/(qnorm(0.75)-qnorm(0.25)),
                  betaGE_SNP_norm=c(log(1.3), 0, 0, log(1.3), 0) / (qnorm(0.75)-qnorm(0.25)),
                  MAF=c(0.1, 0.3, 0.3, 0.3, 0.1),
                  SNP_cor=0.7,
                  regress_E_norm_on_G_SNP=c(sqrt(0.001),rep(0,4)),
                  control=list(trace=1))
}
\seealso{
\code{\link{spmleCombo}}, \code{\link{spmle}}
}
