% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_functions.R
\name{spmleCombo}
\alias{spmleCombo}
\title{Improved Semiparametric Estimator for Case-Control Studies Under G-E Independence.}
\usage{
spmleCombo(D, G, E, pi1, data, nboot = 50, ncores = 1, control = list(),
  startvals)
}
\arguments{
\item{D}{a binary vector of disease status (1=case, 0=control).}

\item{G}{a vector or matrix (if multivariate) containing genetic data. Can be continuous, discrete, or a combination.}

\item{E}{a vector or matrix (if multivariate) containing environmental data. Can be continuous, discrete, or a combination.}

\item{pi1}{the population disease rate, a scalar in [0, 1) or the string "rare".
Using \code{pi1=0} is the rare disease approximation.}

\item{data}{an optional list or environment containing the variables in the model.
If not found in \code{data}, the variables are taken from the environment
from which \code{spmleCombo} is called.}

\item{nboot}{the number of bootstraps to use when estimating the
standard error, an integer.  Setting \code{nboot=0} disables the bootstrap
and uses the asymptotic standard error estimate (not recommended because the
asymptotic SE often has poor coverage: setting \code{nboot=0} will
trigger a warning).  Default \code{nboot = 50}.}

\item{ncores}{the number of cpu cores to use when parallelizing bootstraps, an integer.
Default \code{ncores = 1} executes the bootstrap sequentially.}

\item{control}{a list of control parameters that allow the user to control the optimization algorithm.  See 'Details'.}

\item{startvals}{an optional numeric vector of coefficient starting values for optimization.  Usually left blank,
in which case logistic regression estimates are used as starting values.}
}
\value{
an object of class \code{"spmle"}.  The function \code{summary} (i.e., \code{summary.spmle})
  can be used to obtain or print a summary of the results.  The Symmetric Combination Estimator
  is not a maximum (pseudo)likelihood estimator like \code{\link{spmle}}; it is the
  optimal combination of two such estimators.  As such, it has no associated loglikelihood
  and the function \code{anova.spmle} cannot be used to compare models fit using
  \code{spmleCombo}.

  \code{\link{predict.spmle}}, the \code{predict} method for S3 class \code{"spmle"},
  can predict the expected response (on logistic or probability scales), compute
  confidence intervals for the expected response, and provide standard errors.

  The generic accessor functions \code{coefficients}, \code{fitted.values}
  and \code{residuals} can be used to extract various useful features of the
  value returned by \code{spmleCombo}.

  An object of class \code{"spmle"} is a list containing at least the following components:
\describe{
  \item{\code{coefficients}}{a named vector of coefficients}
  \item{\code{pi1}}{the value of pi1 used during the analysis}
  \item{\code{SE}}{standard error estimate of coefficients}
  \item{\code{cov}}{estimated covariance matrix of coefficients}
  \item{\code{glm_fit}}{a logistic regression model fit using the same model as \code{spmleCombo}}
  \item{\code{call}}{the matched call}
  \item{\code{formula}}{the formula used}
  \item{\code{data}}{the \code{data argument}}
  \item{\code{model}}{the model frame}
  \item{\code{terms}}{the \code{terms} object used}
  \item{\code{linear.predictors}}{the linear fit on the logistic link scale}
  \item{\code{fitted.values}}{the fitted values on the probability scale}
  \item{\code{residuals}}{the Pearson residuals}
  \item{\code{null.deviance}}{the deviance for the null model}
  \item{\code{df.residual}}{the residual degrees of freedom}
  \item{\code{df.null}}{the residual degrees of freedom for the null model}
  \item{\code{rank}}{the numeric rank of the fitted linear model
    (i.e. the number of parameters estimated)}
  \item{\code{nobs}}{number of observations}
  \item{\code{ncase}}{number of cases}
  \item{\code{ncontrol}}{number of controls}
}

\code{spmle} objects created by \code{spmleCombo()} additionally have components \code{spmle_E}
(model from \code{spmle} that profiled out the distribution of E), \code{spmle_G}
(model from \code{spmle} that profiled out the distribution of G with \code{swap=TRUE}),
and \code{bootstraps} (matrix of bootstrapped parameter estimates, if \code{nboot > 0}).
}
\description{
\code{spmleCombo} estimates Gene-Environment interactions in case-control data under the assumption
of G-E independence in the underlying population.  This is an improved version of
\code{\link{spmle}} that, on average, has significantly smaller mean squared error
than \code{spmle}, at the cost of increased computing time.
}
\details{
This function calculates the Symmetric Combination Estimator of Wang et. al. (2018),
which improves estimation efficiency in case-control studies of gene-environment
interactions while treating the marginal distributions of G and E nonparametrically.

This is done by calling \code{spmle} twice (once with \code{swap=TRUE}) to generate
two symmetric estimates, which are then combined using a GLS approach.
It currently supports the model with G and E main effects and a multiplicative G*E interaction.

The \code{control} argument is a list that controls the behavior of the optimization algorithm
\code{\link[ucminf]{ucminf}} from the \pkg{ucminf} package.  When \code{ucminf} works,
it works brilliantly (typically more than twice as fast as the next-fastest algorithm).
But it has a nasty habit of declaring convergence before actually converging.
To address this, \code{spmleCombo} checks the maximum gradient at "convergence", and can rerun the optimization
using different starting values.  The \code{control} argument can supply any of the following components:
\describe{
  \item{\code{max_grad_tol}}{maximum allowable gradient at convergence.  \code{spmleCombo} does not
    consider the optimization to have converged if the maximum gradient \code{> max_grad_tol}
    when \code{ucminf} stops. Default \code{max_grad_tol} \code{= 0.001}.}
  \item{\code{num_retries}}{number of times to retry optimization.  An error is produced if
    the optimization has not converged after \code{num_retries}.  Different starting values
    are used for each retry.  Default \code{num_retries = 2}.}
  \item{\code{use_hess}}{a logical value instructing \code{spmleCombo} to use the analytic hessian
    to precondition the optimization.  This brings significant speed benefits, and is one reason
    \code{ucminf} is so fast.  For unknown reasons, preconditioning causes computers with
    certain Intel CPUs to prematurely terminate iterating.  By default, \code{use_hess = TRUE},
    but if you notice that \code{ucminf} never converges during the first attempt, try setting \code{use_hess = FALSE}.}
  \item{\code{trace}}{a scalar or logical value that is used by both \code{spmleCombo} and \code{ucminf}
    to control the printing of detailed tracing information.
    If TRUE or > 0, details of each \code{ucminf} iteration are printed.
    If FALSE or 0, \code{ucminf} iteration details are suppressed but \code{spmleCombo} still
    prints optimization retries.  If \code{trace < 0} nothing is printed.  Default \code{trace = 0}.}
  \item{additional control parameters}{not used by \code{spmleCombo}, but are passed to \code{\link[ucminf]{ucminf}}.
    Note that the \code{ucminf} algorithm has four stopping criteria, and \code{ucminf} will
    declare convergence if any one of them has been met.  The \code{ucminf} control parameter
    "\code{grtol}" controls \code{ucminf}'s gradient stopping criterion, which defaults to
    \code{1e-6}.  \code{grtol} should not be set larger than the \code{spmleCombo} control parameter \code{max_grad_tol}.}
}
}
\section{References}{

Stalder, O., Asher, A., Liang, L., Carroll, R. J., Ma, Y., and Chatterjee, N. (2017).
\emph{Semi-parametric analysis of complex polygenic gene-environment interactions in case-control studies.}
Biometrika, 104, 801–812.

Wang, T., Asher, A., Carroll, R. J. (2018).
\emph{Improved Semiparametric Analysis of Polygenic Gene-Environment Interactions in Case-Control Studies}
Unpublished.
}

\examples{
# Simulation from Table 1 in Stalder et. al. (2017)
set.seed(2018)
dat = simulateCC(ncase=500, ncontrol=500, beta0=-4.165,
                 betaG_SNP=c(log(1.2), log(1.2), 0, log(1.2), 0),
                 betaE_bin=log(1.5),
                 betaGE_SNP_bin=c(log(1.3), 0, 0, log(1.3), 0),
                 MAF=c(0.1, 0.3, 0.3, 0.3, 0.1),
                 SNP_cor=0.7, E_bin_freq=0.5)

# SPMLE with known population disease rate of 0.03 and asymptotic SE estimates
spmleCombo(D=D, G=G, E=E, pi1=0.03, data=dat, nboot=0)
}
\seealso{
\code{\link{spmle}}, \code{\link{simulateCC}} to simulate data
}
