% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{string2num}
\alias{string2num}
\title{Fail-tolerant function to convert a string to numeric if applicable}
\usage{
string2num(string)
}
\arguments{
\item{string}{string with potential numeric values}
}
\value{
Numeric variable (if appropriate), otherwise string
}
\description{
Fail-tolerant function to convert a string to numeric if applicable
}
