% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{trimwhite}
\alias{trimwhite}
\title{Function to trim leading and trailing whitespace}
\usage{
trimwhite(x)
}
\arguments{
\item{x}{string to be trimmed}
}
\value{
Trimmed string
}
\description{
Function to trim leading and trailing whitespace
}
