% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_functions.R
\name{combo_asymp}
\alias{combo_asymp}
\title{Calculate seven estimators with asymptotic SEs: logistic, SPMLE, swap, Optimal Combo, composite SPMLE}
\usage{
combo_asymp(D, G, E, pi1, control = list())
}
\arguments{
\item{D}{a binary vector of disease status (1=case, 0=control).}

\item{G}{a vector or matrix (if multivariate) containing genetic data. Can be continuous, discrete, or a combination.}

\item{E}{a vector or matrix (if multivariate) containing environmental data. Can be continuous, discrete, or a combination.}

\item{pi1}{the population disease rate, a scalar in [0, 1).  Using \code{pi1=0} is the rare disease approximation.}

\item{control}{a list of control parameters.
\code{nboot} number of bootstraps.  Default \code{0}.
\code{trace} is a scalar.  If >-1, tracing information is produced.  Default \code{0}.
\code{usehess} logical: precondition optimization with the hessian.  Default \code{FALSE}.}
}
\value{
A list with a matrix of estimates and SEs and a bunch of other stuff
}
\description{
Calculate seven estimators with asymptotic SEs: logistic, SPMLE, swap, Optimal Combo, composite SPMLE
}
\seealso{
\code{\link{simulate_complex}} to simulate data
}
