% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateCC.R
\name{enn}
\alias{enn}
\title{Checks whether a variable exists (in the current environment) and is not NULL}
\usage{
enn(varname, n = 1)
}
\arguments{
\item{varname}{variable name}
}
\value{
TRUE or FALSE
}
\description{
Checks whether a variable exists (in the current environment) and is not NULL
}
