% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.spmle}
\alias{predict.spmle}
\title{Predict method for spmle objects}
\usage{
\method{predict}{spmle}(object, newdata, se.fit = FALSE,
  interval = c("none", "confidence"), level = 0.95, type = c("link",
  "response"), na.action = na.pass, ...)
}
\arguments{
\item{object}{of class inheriting from "spmle"}

\item{newdata}{an optional list or data frame in which to look for variables
to use when making predictions. If omitted, the fitted values are used.}

\item{se.fit}{a switch indicating if standard errors are required.}

\item{interval}{Type of interval calculation. Can be abbreviated. Prediction
intervals are not meaningful for binary responses and are not allowed.}

\item{level}{confidence level.}

\item{type}{the type of prediction required. The default is \code{"link"} which
uses the logit scale of the linear predictors, giving the log odds.
The alternative \code{"response"} uses the probability scale, giving
\code{Pr(D=1|G,E)}.}

\item{na.action}{function determining what should be done with missing values
in newdata. The default is to predict NA.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
\code{predict.spmle} produces a vector of predictions or a matrix of
  predictions and bounds with column names fit, lwr, and upr if interval is set.

  If \code{se.fit} is \code{TRUE}, a list with the following components is returned:
  \describe{
    \item{\code{fit}}{vector or matrix as above}
    \item{\code{se.fit}}{vector with standard error of predicted means, in the link scale}
    \item{\code{residual.scale}}{residual standard deviation}
    \item{\code{df}}{degrees of freedom for residual}
  }
}
\description{
Obtains predictions from a fitted spmle object.
}
\details{
\code{predict.spmle} produces predicted values, obtained by evaluating the
\code{spmle} function in the frame newdata (which defaults to \code{model.frame(object)}).
If the logical \code{se.fit} is \code{TRUE}, standard errors of the predictions
are calculated (only in the link scale). Setting \code{interval="confidence"} specifies computation of
confidence intervals at the specified level.
}
